requirements.txt
streamlit==1.24.0
python-docx==0.8.11
pdfminer.six==20221105
spacy==3.5.3
torch==2.0.1
transformers==4.30.2
pandas==2.0.3
scikit-learn==1.3.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0.tar.gz
pyresparser==1.0.6
python-multipart==0.0.6